requirements.txt
asgiref==3.9.1
mercadopago==2.3.0
Django==5.2.6
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
gunicorn
psycopg2-binary
dj-database-url==2.1.0
python-dotenv==1.1.1
